refactoring types

+++ libs/data-storage/!data-row-accessor.h
@@ -6,7 +6,7 @@ class DataFieldAccessor { // aka DataRowWrapper
 public:
 
     virtual size_t FieldsCount() = 0;
-    virtual FieldData GetField(int field_num) = 0;
+    virtual FieldData GetField(size_t field_num) = 0;
     virtual FieldData GetField(const char* field_name) = 0;
     virtual std::string_view GetFieldName(size_t field_num) = 0;
     virtual DataRow GenRow() = 0;
